Fix bug where get logs by block number range fails
This PR is an attempt at solving an issue where
retrieving logs by a range that is define using
block numbers, fails.

This seems to be because the
`DefaultBlockParameterNumber` instance
is serialized to JSON incorrectly.
i.e. `DefaultBlockParameterNumber` is serialized
as an object with a `value` attribute instead of
as a JSON string.